handler/requirements.txt
python-keycloak @ git+https://github.com/marcospereirampj/python-keycloak@a63982188ae14e1a0948f324bb3e9a8387650e65
SQLAlchemy==1.4.23
psycopg2-binary==2.9.5
